+++ fcli-core/fcli-common/src/main/java/com/fortify/cli/common/http/ssl/truststore/helper/TrustStoreConfigDescriptor.java
@@ -27,4 +27,5 @@ public class TrustStoreConfigDescriptor extends JsonNodeHolder {
     private String path;
     private String type;
     private String password;
+    private Boolean useOsTrustStore;
 }